requirement_pip_list.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.8
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-translate==3.20.1
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.3
openai==1.64.0
outcome==1.3.0.post0
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
PyMySQL==1.1.1
pyparsing==3.2.1
pypdfium2==4.30.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
threadpoolctl==3.5.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
